Time: 13 ms (49.14%), Space: 17.3 MB (48.00%) - LeetHub

+++ 662-maximum-width-of-binary-tree/662-maximum-width-of-binary-tree.cpp
@@ -13,17 +13,17 @@ class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
         if(root == NULL) return 0;
-        int ans =0;
-        queue<pair<TreeNode*,int>> q;
+        unsigned long long int ans =0;
+        queue<pair<TreeNode*,unsigned long long int>> q;
         q.push({root,0});
         
         while(!q.empty()){
             int n = q.size();
-            int mi = q.front().second;
-            int last,first;
+            unsigned long long int mi = q.front().second;
+            unsigned long long int last,first;
             
             for(int i=0;i<n;i++){
-                int curr_idx = q.front().second-mi;
+                unsigned long long int curr_idx = q.front().second-mi;
                 TreeNode* node = q.front().first;
                 q.pop();
                 if(i==0) first = curr_idx;
